Updated actions to point to their combined counterparts, refactored LTR slightly to use hooks
