+++ src/applications/debt-letters/components/DebtLettersWrapper.jsx
@@ -1,10 +1,9 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
-import { connect } from 'react-redux';
-import { bindActionCreators } from 'redux';
+import { connect, useDispatch } from 'react-redux';
 import FEATURE_FLAG_NAMES from 'platform/utilities/feature-toggles/featureFlagNames';
 import { toggleValues } from 'platform/site-wide/feature-toggles/selectors';
-import { fetchDebtLetters } from '../actions';
+import { fetchDebtLetters } from '../../combined-debt-portal/combined/actions/debts';
 
 const DebtLettersWrapper = ({
   isPending,
@@ -13,15 +12,17 @@ const DebtLettersWrapper = ({
   showDebtLetters,
   isProfileUpdating,
   isLoggedIn,
-  getDebtLetters,
 }) => {
+  const dispatch = useDispatch();
+
   useEffect(
     () => {
       if (showDebtLetters) {
-        getDebtLetters();
+        const generateGetDebtLetters = () => fetchDebtLetters(dispatch);
+        dispatch(generateGetDebtLetters);
       }
     },
-    [getDebtLetters, showDebtLetters],
+    [dispatch, showDebtLetters],
   );
 
   if (isPending || isPendingVBMS || isProfileUpdating) {
@@ -66,13 +67,8 @@ const mapStateToProps = state => ({
   ],
 });
 
-const mapDispatchToProps = dispatch => ({
-  ...bindActionCreators({ getDebtLetters: fetchDebtLetters }, dispatch),
-});
-
 DebtLettersWrapper.propTypes = {
   children: PropTypes.array,
-  getDebtLetters: PropTypes.func,
   isLoggedIn: PropTypes.bool,
   isPending: PropTypes.bool,
   isPendingVBMS: PropTypes.bool,
@@ -85,7 +81,4 @@ DebtLettersWrapper.defaultProps = {
   isPendingVBMS: false,
 };
 
-export default connect(
-  mapStateToProps,
-  mapDispatchToProps,
-)(DebtLettersWrapper);
+export default connect(mapStateToProps)(DebtLettersWrapper);
